requirements.txt
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==2.3.2
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-core==2.4.1
google-cloud-firestore==2.18.0
google-cloud-secret-manager==2.20.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.1
grpcio-status==1.66.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
requests==2.31.0
rsa==4.9
urllib3==2.2.3
Werkzeug==3.0.4
